backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
nltk==3.8.2
numpy==1.26.4
openai==1.40.6
pandas==2.2.1
pdfminer==20191125
pdfminer.six==20240706
PyPDF2==3.0.1
python-dotenv==1.0.1
setuptools==68.2.2
spacy==3.7.5
typing_extensions==4.12.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl